Here is a version with the file manager doing some encode and decode from an example provided by vtabmow at this thread. https://www.hackingwithswift.com/forums/100-days-of-swiftui/day-68-documents-directory-challenge/13319
I had to create another variable in the method declaration , dataType: T to explicitly define what T was because it just would not infer it through using a line like "let input : String =" in the call. It works now, but it has quite a few warnings that don't look pleasant.